+++ test/libsolidity/syntaxTests/inheritance/allow_empty_duplicated_super_constructor_call.sol
@@ -1,0 +1,3 @@
+contract A { constructor() public { } }
+contract B1 is A { constructor() A() public {  } }
+contract B2 is A { constructor() A public {  } }