+++ test/libsolidity/syntaxTests/inheritance/duplicated_constructor_call/base.sol
@@ -1,0 +1,4 @@
+contract A { constructor(uint) public { } }
+contract B is A(2) { constructor() A(3) public {  } }
+// ----
+// Warning: Base constructor arguments given twice.