+++ test/libsolidity/syntaxTests/inheritance/duplicated_constructor_call/base_V050.sol
@@ -1,0 +1,6 @@
+pragma experimental "v0.5.0";
+
+contract A { constructor(uint) public { } }
+contract B is A(2) { constructor() A(3) public {  } }
+// ----
+// DeclarationError: Base constructor arguments given twice.